requirements.txt
requests
Pillow
streamlit
opencv-python
schedule